Prueba E

+++ main.go
@@ -3,7 +3,7 @@ package main
 import "fmt"
 
 func main() {
-	var a, b float64
+	var a, b int64
 	fmt.Scan(&a)
 	b = a * a
 	fmt.Println(b)
